+++ reactivesocket-netty/src/main/java/io/reactivesocket/netty/NettyDuplexConnection.java
@@ -13,80 +13,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.reactivesocket.netty.tcp.client;
+package io.reactivesocket.netty;
 
-import io.netty.bootstrap.Bootstrap;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import io.netty.channel.*;
-import io.netty.channel.socket.SocketChannel;
-import io.netty.channel.socket.nio.NioSocketChannel;
-import io.netty.handler.codec.LengthFieldBasedFrameDecoder;
 import io.reactivesocket.DuplexConnection;
 import io.reactivesocket.Frame;
 import io.reactivesocket.exceptions.TransportException;
 import io.reactivesocket.rx.Completable;
 import io.reactivesocket.rx.Observable;
 import io.reactivesocket.rx.Observer;
-import org.agrona.BitUtil;
 import org.reactivestreams.Publisher;
 import org.reactivestreams.Subscriber;
 import org.reactivestreams.Subscription;
 
 import java.io.IOException;
-import java.net.SocketAddress;
 import java.nio.channels.ClosedChannelException;
 import java.util.concurrent.CopyOnWriteArrayList;
 
-public class ClientTcpDuplexConnection implements DuplexConnection {
-    private final Channel channel;
-    private final CopyOnWriteArrayList<Observer<Frame>> subjects;
+public abstract class NettyDuplexConnection implements DuplexConnection {
+    protected final Channel channel;
+    protected final CopyOnWriteArrayList<Observer<Frame>> readers;
 
-    private ClientTcpDuplexConnection(Channel channel, CopyOnWriteArrayList<Observer<Frame>> subjects) {
-        this.subjects  = subjects;
+    protected NettyDuplexConnection(Channel channel, CopyOnWriteArrayList<Observer<Frame>> readers) {
+        this.readers = readers;
         this.channel = channel;
-    }
-
-    public static Publisher<ClientTcpDuplexConnection> create(SocketAddress address, EventLoopGroup eventLoopGroup) {
-        return s -> {
-            CopyOnWriteArrayList<Observer<Frame>> subjects = new CopyOnWriteArrayList<>();
-            ReactiveSocketClientHandler clientHandler = new ReactiveSocketClientHandler(subjects);
-            Bootstrap bootstrap = new Bootstrap();
-            ChannelFuture connect = bootstrap
-                .group(eventLoopGroup)
-                .channel(NioSocketChannel.class)
-                .option(ChannelOption.TCP_NODELAY, true)
-                .option(ChannelOption.SO_REUSEADDR, true)
-                .option(ChannelOption.AUTO_READ, true)
-                .option(ChannelOption.CONNECT_TIMEOUT_MILLIS, 10000)
-                .handler(new ChannelInitializer<SocketChannel>() {
-                    @Override
-                    protected void initChannel(SocketChannel ch) throws Exception {
-                        ChannelPipeline p = ch.pipeline();
-                        p.addLast(
-                            new LengthFieldBasedFrameDecoder(Integer.MAX_VALUE >> 1, 0, BitUtil.SIZE_OF_INT, -1 * BitUtil.SIZE_OF_INT, 0),
-                            clientHandler
-                        );
-                    }
-                }).connect(address);
-
-            connect.addListener(connectFuture -> {
-                if (connectFuture.isSuccess()) {
-                    Channel ch = connect.channel();
-                    s.onNext(new ClientTcpDuplexConnection(ch, subjects));
-                    s.onComplete();
-                } else {
-                    s.onError(connectFuture.cause());
-                }
-            });
-        };
     }
 
     @Override
     public final Observable<Frame> getInput() {
         return o -> {
-            o.onSubscribe(() -> subjects.removeIf(s -> s == o));
-            subjects.add(o);
+            o.onSubscribe(() -> readers.removeIf(s -> s == o));
+            readers.add(o);
         };
     }
 
@@ -144,21 +103,5 @@ public class ClientTcpDuplexConnection implements DuplexConnection {
     @Override
     public void close() throws IOException {
         channel.close();
-    }
-
-    public String toString() {
-        if (channel == null) {
-            return "ClientTcpDuplexConnection(channel=null)";
-        }
-
-        return "ClientTcpDuplexConnection(channel=[" +
-            "remoteAddress=" + channel.remoteAddress() + "," +
-            "isActive=" + channel.isActive() + "," +
-            "isOpen=" + channel.isOpen() + "," +
-            "isRegistered=" + channel.isRegistered() + "," +
-            "isWritable=" + channel.isWritable() + "," +
-            "channelId=" + channel.id().asLongText() +
-            "])";
-
     }
 }
