+++ reactivesocket-transport-tcp/src/main/java/io/reactivesocket/transport/tcp/client/ReactiveSocketClientHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.reactivesocket.netty.tcp.client;
+package io.reactivesocket.transport.tcp.client;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelHandler;
