+++ reactivesocket-transport-websocket/build.gradle
@@ -1,0 +1,6 @@
+dependencies {
+    compile project(':reactivesocket-core')
+    compile project(':reactivesocket-netty')
+
+    testCompile project(':reactivesocket-test')
+}
