+++ reactivesocket-transport-websocket/src/main/java/io/reactivesocket/transport/websocket/client/WebSocketReactiveSocketConnector.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.reactivesocket.netty.websocket.client;
+package io.reactivesocket.transport.websocket.client;
 
 import io.netty.channel.EventLoopGroup;
 import io.reactivesocket.*;
