+++ settings.gradle
@@ -1,7 +1,9 @@
 rootProject.name='reactivesocket'
 include 'reactivesocket-core'
-include 'reactivesocket-transport-aeron'
-include 'reactivesocket-transport-netty'
-include 'reactivesocket-transport-local'
+include 'reactivesocket-netty'
 include 'reactivesocket-mime-types'
 include 'reactivesocket-test'
+include 'reactivesocket-transport-aeron'
+include 'reactivesocket-transport-local'
+include 'reactivesocket-transport-tcp'
+include 'reactivesocket-transport-websocket'
